feat: impl execute task (#23)
* feat: impl execute task

* feat: cache api key

* feat: query execution and task version by task_id

+++ tokenspan-api/src/api/parameter/dto/parameter_input.rs
@@ -5,7 +5,7 @@ use async_graphql::InputObject;
 pub struct ParameterCreateInput {
     pub name: String,
     pub temperature: f32,
-    pub max_tokens: u32,
+    pub max_tokens: u16,
     pub stop_sequences: Vec<String>,
     pub top_p: f32,
     pub frequency_penalty: f32,
